+++ selfdrive/ui/translations/main_ja.ts
@@ -367,9 +367,9 @@
     </message>
     <message>
         <source>Get turn-by-turn directions displayed and more with a comma
-prime subscription. Sign up now: https://connect.comma.ai</source>
+prime subscription. Sign up now: https://connect.retropilot.org</source>
         <translation>より詳細な案内情報を得ることができます。
-詳しくはこちら：https://connect.comma.ai</translation>
+詳しくはこちら：https://connect.retropilot.org</translation>
     </message>
     <message>
         <source>No home
@@ -451,16 +451,16 @@ location set</source>
         <translation>デバイスと comma アカウントを連携する</translation>
     </message>
     <message>
-        <source>Go to https://connect.comma.ai on your phone</source>
-        <translation>スマートフォンで「https://connect.comma.ai」にアクセスしてください。</translation>
+        <source>Go to https://connect.retropilot.org on your phone</source>
+        <translation>スマートフォンで「https://connect.retropilot.org」にアクセスしてください。</translation>
     </message>
     <message>
         <source>Click &quot;add new device&quot; and scan the QR code on the right</source>
         <translation>「新しいデバイスを追加」を押し、右側のQRコードをスキャンしてください。</translation>
     </message>
     <message>
-        <source>Bookmark connect.comma.ai to your home screen to use it like an app</source>
-        <translation>「connect.comma.ai」をホーム画面に追加して、アプリのように使うことができます。</translation>
+        <source>Bookmark connect.retropilot.org to your home screen to use it like an app</source>
+        <translation>「connect.retropilot.org」をホーム画面に追加して、アプリのように使うことができます。</translation>
     </message>
 </context>
 <context>
@@ -470,8 +470,8 @@ location set</source>
         <translation>今すぐｱｯﾌﾟｸﾞﾚｰﾄ</translation>
     </message>
     <message>
-        <source>Become a comma prime member at connect.comma.ai</source>
-        <translation>connect.comma.ai でﾌﾟﾗｲﾑ会員に登録できます</translation>
+        <source>Become a comma prime member at connect.retropilot.org</source>
+        <translation>connect.retropilot.org でﾌﾟﾗｲﾑ会員に登録できます</translation>
     </message>
     <message>
         <source>PRIME FEATURES:</source>
@@ -501,8 +501,8 @@ location set</source>
         <translation>comma prime</translation>
     </message>
     <message>
-        <source>CONNECT.COMMA.AI</source>
-        <translation>CONNECT.COMMA.AI</translation>
+        <source>CONNECT.retropilot.org</source>
+        <translation>CONNECT.retropilot.org</translation>
     </message>
     <message>
         <source>COMMA POINTS</source>
@@ -709,8 +709,8 @@ location set</source>
         <translation>セットアップ完了</translation>
     </message>
     <message>
-        <source>Pair your device with comma connect (connect.comma.ai) and claim your comma prime offer.</source>
-        <translation>デバイスを comma connect (connect.comma.ai)でペアリングし、comma primeの特典を申請してください。</translation>
+        <source>Pair your device with comma connect (connect.retropilot.org) and claim your comma prime offer.</source>
+        <translation>デバイスを comma connect (connect.retropilot.org)でペアリングし、comma primeの特典を申請してください。</translation>
     </message>
     <message>
         <source>Pair device</source>